verification: Add auto-merge step
The step enables auto-merge on a PR if the following conditions
are fulfilled:
- author of the PR is among trusted authors (defined in this workflow)
- 'results' label is present (workflow adds the label itself)

This commit also makes the action to give a positive review to the PR
if the uploaded files pass verification.
Auto-marge must be enable both on the repo and organization settings
level. Auto-merge can only happen if all the branch protection rules
are fulfilled.
(required checks passed, amount of positive reviews, etc.)

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>